bmw/requirements.txt
ansible==5.0.1
ansible-core==2.12.1
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==36.0.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.6
django-background-tasks==1.2.5
django-compat==1.0.15
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
future==0.18.2
html5==0.0.9
idna==3.3
ipykernel==6.10.0
ipython==8.2.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
kiwisolver==1.4.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.21.4
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipfile==0.0.2
plotly==5.6.0
prometheus-client==0.12.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.28.1
resolvelib==0.5.4
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
sqlitedict==2.0.0
sqlparse==0.4.2
stack-data==0.2.0
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
